feat: outsource search
